+++ PROEVENTOS/Front/ProEventos-App/src/app/components/user/perfil/perfil-detalhe/perfil-detalhe.component.spec.ts
@@ -1,0 +1,28 @@
+/* tslint:disable:no-unused-variable */
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
+
+import { PerfilDetalheComponent } from './perfil-detalhe.component';
+
+describe('PerfilDetalheComponent', () => {
+  let component: PerfilDetalheComponent;
+  let fixture: ComponentFixture<PerfilDetalheComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ PerfilDetalheComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(PerfilDetalheComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
