+++ PROEVENTOS/Front/ProEventos-App/src/app/components/user/perfil/perfil.component.ts
@@ -13,91 +13,18 @@ import { ValidatorField } from '@app/helpers/ValidatorField';
   styleUrls: ['./perfil.component.scss'],
 })
 export class PerfilComponent implements OnInit {
-  userUpdate = {} as UserUpdate;
-  form!: FormGroup;
+
 
   constructor(
-    private fb: FormBuilder,
-    public accountService: AccountService,
-    private router: Router,
-    private toaster: ToastrService,
-    private spinner: NgxSpinnerService
+
   ) { }
 
   ngOnInit(): void {
-    this.validation();
-    this.carregarUsuario();
+
   }
 
-  private carregarUsuario(): void {
-    this.spinner.show();
-    this.accountService
-      .getUser()
-      .subscribe(
-        (userRetorno: UserUpdate) => {
-          console.log(userRetorno);
-          this.userUpdate = userRetorno;
-          this.form.patchValue(this.userUpdate);
-          this.toaster.success('Usuário Carregado', 'Sucesso');
-        },
-        (error) => {
-          console.error(error);
-          this.toaster.error('Usuário não Carregado', 'Erro');
-          this.router.navigate(['/dashboard']);
-        }
-      )
-      .add(() => this.spinner.hide());
-  }
-
-  private validation(): void {
-    const formOptions: AbstractControlOptions = {
-      validators: ValidatorField.MustMatch('password', 'confirmePassword'),
-    };
-
-    this.form = this.fb.group(
-      {
-        userName: [''],
-        titulo: ['NaoInformado', Validators.required],
-        primeiroNome: ['', Validators.required],
-        ultimoNome: ['', Validators.required],
-        email: ['', [Validators.required, Validators.email]],
-        phoneNumber: ['', [Validators.required]],
-        descricao: ['', Validators.required],
-        funcao: ['NaoInformado', Validators.required],
-        password: ['', [Validators.minLength(4), Validators.nullValidator]],
-        confirmePassword: ['', Validators.nullValidator],
-      },
-      formOptions
-    );
-  }
-
-  // Conveniente para pegar um FormField apenas com a letra F
   get f(): any {
-    return this.form.controls;
+    return '';
   }
 
-  onSubmit(): void {
-    this.atualizarUsuario();
-  }
-
-  public atualizarUsuario() {
-    this.userUpdate = { ...this.form.value };
-    this.spinner.show();
-
-    this.accountService
-      .updateUser(this.userUpdate)
-      .subscribe(
-        () => this.toaster.success('Usuário atualizado!', 'Sucesso'),
-        (error) => {
-          this.toaster.error(error.error);
-          console.error(error);
-        },
-      )
-      .add(() => this.spinner.hide());
-  }
-
-  public resetForm(event: any): void {
-    event.preventDefault();
-    this.form.reset();
-  }
 }
